Fix libspec source modifier
